requirements.txt
scrapy
pycrypto
pymongo